web-app/requirements.txt
Flask

Flask-WTF

Flask-Bootstrap4

pickle-mixin

nltk

bs4

regex

unicodedata2

langdetect

deep_translator

sklearn

numpy

gunicorn

pandas